% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{add_job}
\alias{add_job}
\title{Run Next Task In Bettr Task Queue}
\usage{
add_job(bettr_tasks)
}
\arguments{
\item{bettr_tasks}{The tasks to be added to the job
table.  Refer to the bettr readme for information
about specifying bettr tasks for this function.}
}
\description{
Checks the next runnable task in the bettr task table
and runs it.  If no eligible task is found, this logic
will return with no side-effects.
Adds A Job to the Bettr Host Task Table
}
