% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{add_job_to_host}
\alias{add_job_to_host}
\title{Adds A Job to the Bettr Host Task Table}
\usage{
add_job_to_host(bettr_tasks)
}
\arguments{
\item{bettr_tasks}{The tasks to be added to the job
table.  Refer to \code{?bettr::get_bettr_task_defaults}
for information about bettr_task columns.}
}
\description{
Adds A Job to the Bettr Host Task Table
}
