% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{append_rows}
\alias{append_rows}
\title{Append Rows to a Table}
\usage{
append_rows(rows, connection_name, table_name, suppress_bind_logging = FALSE)
}
\arguments{
\item{rows}{a data frame, data.table, or tibble with rows of values to insert}

\item{connection_name}{The snake_case name of the connection.}

\item{table_name}{The name of the table to receive data. Any occurrences of
\verb{@@} in the table name will be replaced by the contents of the environment
variable \code{BETTR_AT_SCOPE} if it contains only alpha characters.}

\item{suppress_bind_logging}{Optionally suppress the logging of bind parameters (defaults to false).}
}
\description{
Take the passed rows and append them to a table given a connection.
If the table does not exist, it is created with \code{ensure_table()}.
Note:
\itemize{
\item The schema for created tables is expected to be the database username.
\item Rows are appended in batches (strides) of 10,000 to avoid buffer overflow.
}
}
