% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{assert_valid_bettr_tasks}
\alias{assert_valid_bettr_tasks}
\title{Asserts a Bettr Task Has Needed Columns}
\usage{
assert_valid_bettr_tasks(bettr_tasks = tibble::tibble(), cols = c())
}
\description{
Asserts a Bettr Task Has Needed Columns
}
