% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{close_db_conn}
\alias{close_db_conn}
\title{Close a Database Connection}
\usage{
close_db_conn(conn)
}
\arguments{
\item{conn}{The connection object to close.}
}
\value{
Response from DBI::dbDisconnect
}
\description{
This function closes a database connection.
}
