% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{drop_table}
\alias{drop_table}
\title{Drop a Table}
\usage{
drop_table(connection_name, table_name)
}
\arguments{
\item{connection_name}{The snake_case name of the connection.}

\item{table_name}{The snake_case name of the table to drop.}
}
\description{
Drops a table from the database.
Note that this function depends on a SQL file named \verb{drop_table_<table_name>.sql}
in the connection-specific directory.
}
