% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{ensure_table}
\alias{ensure_table}
\title{Ensure a Table Exists for Appending Rows}
\usage{
ensure_table(connection_name, table_name)
}
\arguments{
\item{connection_name}{The snake_case name of the connection.}

\item{table_name}{The snake_case name of the table to ensure exists in the database.}
}
\value{
Nothing
}
\description{
This function ensures a table exists in the named connection for appending rows.
Note that this function depends on a SQL file named \verb{create_table_<table_name>.sql}
in the connection-specific directory.
}
