% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{ensure_view}
\alias{ensure_view}
\title{Ensure a View Exists}
\usage{
ensure_view(connection_name, view_name)
}
\arguments{
\item{connection_name}{The snake_case name of the connection.}

\item{view_name}{The snake_case name of the view to ensure
exists in the database.}
}
\value{
TRUE if view exists or was initialized successfully.
}
\description{
Calling this function will check to see if a view exists
at the connection with the given name.
If it does not exist, bettr will look for an initialization
file in \verb{SQL_DIR/<CONNECTION_NAME>/init_<view_name>.sql}.
}
