% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bettr_data.R
\name{ets_model}
\alias{ets_model}
\title{ETS model forecasting future betting odds}
\usage{
ets_model(object, odds_vec, h = 36)
}
\arguments{
\item{object}{Object on which \code{ets_model} is performed on.}

\item{odds_vec}{A vector of odds.}

\item{h}{The number of hours to be forecast. Defaults to 36 hours.}
}
\value{
A list featuring the forecasted values, the plot of the odds with respect to time and the model that was fitted
}
\description{
ETS model forecasting future betting odds
}
\examples{
match <- subset(football, football$home_team == "Crystal Palace")
match <- make_bettr(match) \%>\% fill_gaps()
ets_model(match, match$draw_odds, h = 72)
}
