% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{exists_table}
\alias{exists_table}
\title{Check if a Table Exists}
\usage{
exists_table(connection_name, table_name)
}
\arguments{
\item{connection_name}{The snake_case name of the connection.}

\item{table_name}{The snake_case name of the table to check.}
}
\description{
Returns TRUE if a table exists at the given connection by
the given name.
}
