% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{exists_view}
\alias{exists_view}
\title{Check if a View Exists}
\usage{
exists_view(connection_name, view_name)
}
\arguments{
\item{connection_name}{The snake_case name of the connection.}

\item{view_name}{The snake_case name of the view to check.}
}
\description{
Returns TRUE if a view exists at the given connection by
the given name.
}
