% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{flush_to_archive}
\alias{flush_to_archive}
\title{Flush a Table's Stale Rows to Archive}
\usage{
flush_to_archive(connection_name, table_name, stale_column, stale_dt)
}
\arguments{
\item{connection_name}{The snake_case name of the connection.}

\item{table_name}{The name of the (live) table to archive.}

\item{stale_column}{The datetime column used to determine staleness.}

\item{stale_dt}{The datetime threshold used to determine staleness.}
}
\description{
All rows with a \code{stale_column} value less than or equal to
the \code{stale_dt} will be selected, added to the archive table
at \verb{ARCH_<TABLE_NAME>}, and deleted from \verb{<TABLE_NAME>}.
}
