% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{generate_sql_init_union_view_stmt}
\alias{generate_sql_init_union_view_stmt}
\title{Generate ETL Union View Statement}
\usage{
generate_sql_init_union_view_stmt(
  rows,
  table_names,
  view_name,
  connection_name,
  view_grants = c()
)
}
\arguments{
\item{rows}{Rows of data from an extract that are representative of the data
to be stored in the new table and presented in a new view.}

\item{table_names}{Vector of table names to combine in the union.}

\item{view_name}{Name of the union view to create; "V_" prefix will be added.}

\item{connection_name}{The name of the database connection (schema user).}

\item{view_grants}{The names of database roles to grant SELECT permissions.
Defaults to none.}
}
\value{
String containing the generated SQL statement.
}
\description{
This method helps \code{bettr::generate_init_sql} by generating the
statement to create a unioned view of two or more replicate tables.
This is used to create a unioned view of live and archival data.
}
