% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_bind_colnames}
\alias{get_bind_colnames}
\title{Get the Names of Column Binds in a SQL Statement}
\usage{
get_bind_colnames(sql)
}
\arguments{
\item{sql}{The SQL statement to scan for bind parameters.}
}
\value{
A character vector of snake_cased bind parameter names to be used in dplyr verbs.
}
\description{
The query parameters to retrieve data from a database are handled in RORacle as bind parameters.
This function scans a SQL statement for ampersand-prefixed, snake_cased bind parameters.
}
