% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_db_conn}
\alias{get_db_conn}
\title{Get a Database Connection}
\usage{
get_db_conn(connection_name)
}
\arguments{
\item{connection_name}{Database connection name; prefix to env vars.}
}
\value{
A DBI connection object
}
\description{
This function returns a database connection from a simple pool.
We expect each R script invocation to use 1 connection at most per data source.
}
