% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_db_name}
\alias{get_db_name}
\title{Get the Database Connection String or TNS Name}
\usage{
get_db_name(connection_name)
}
\arguments{
\item{connection_name}{the snake_case name of the connection.}
}
\value{
The database connection String as defined in .Renviron
}
\description{
Get the Database Connection String or TNS Name
}
