% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_db_password}
\alias{get_db_password}
\title{Get the Database Password}
\usage{
get_db_password(connection_name)
}
\arguments{
\item{connection_name}{the snake_case name of the connection.}
}
\value{
The database password for that connection as defined in .Renviron
}
\description{
Get the Database Password
}
