% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_db_username}
\alias{get_db_username}
\title{Get the Database Username}
\usage{
get_db_username(connection_name)
}
\arguments{
\item{connection_name}{the snake_case name of the connection.}
}
\value{
The database username for that connection as defined in .Renviron
}
\description{
Get the Database Username
}
