% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_rows_binds}
\alias{get_rows_binds}
\title{Get Rows from a Database With Bind Parameters}
\usage{
get_rows_binds(binds, connection_name, sql, suppress_bind_logging = FALSE)
}
\description{
Calls the SELECT query while passing bind parameters.
Multiple rows of bind parameters are queried in series,
then bound and packaged as a tibble for returning.
Delegate function from get_rows().
}
