% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_rows_nobinds}
\alias{get_rows_nobinds}
\title{Get Rows from a Database Without Bind Parameters}
\usage{
get_rows_nobinds(connection_name, sql)
}
\description{
Calls the SELECT query without passing any bind parameters;
no fancy row-wise manipulation to spread the logic out.
Delegate function from get_rows().
}
