% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_sql_dir}
\alias{get_sql_dir}
\title{Get the Directory where SQL Queries are Stored}
\usage{
get_sql_dir(connection_name = "")
}
\value{
The path to the directory where SQL queries are stored.
}
\description{
All the functions in this module require that SQL statements are stored in external files.
This function refers to the environment variable SQL_DIR and retrieves it for use.
}
