% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_url_query}
\alias{get_url_query}
\title{Get URL Query Parameters}
\usage{
get_url_query(url_query = "")
}
\arguments{
\item{url_query}{A URL query string to parse (optional).  If not passed, session URL query is used.}
}
\value{
Named list of values from URL query
}
\description{
What it says on the tin.  If the session contains a URL query,
this sugar method retrieves it for you.  You can also test by passing
a url_query.
}
