% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{init_task_session}
\alias{init_task_session}
\title{Set Up Task Environment}
\usage{
init_task_session(bettr_task = tibble::tibble(), base_path)
}
\description{
This function loads environment variables and prepares
an options object for use by a task function.
}
