% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{load_sql}
\alias{load_sql}
\title{Load a SQL Statement from an External File}
\usage{
load_sql(sql, connection_name = NA)
}
\arguments{
\item{sql}{The snake_case name of the SQL file without the file extension (such as \code{get_test_data}).}

\item{connection_name}{The snake_case name of the database connection.}
}
\value{
The SQL statement stored in the file.
}
\description{
All the functions in this module require that SQL statements are stored in external files.
This function loads a SQL statement from a file.
}
