% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{open_db_conn}
\alias{open_db_conn}
\title{Open a Database Connection}
\usage{
open_db_conn(connection_name = "micpr")
}
\arguments{
\item{connection_name}{Database connection name; prefix to env vars.}
}
\value{
A DBI connection object
}
\description{
This function opens a database connection.
It depends on environment variables following this convention:
<CONNECTION_NAME>_CREDS_USR: database user name
<CONNECTION_NAME>_CREDS_PSW: database user password
<CONNECTION_NAME>_DB_NAME: database connection string
(tnsnames.ora name OR Oracle connection string)
}
