% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{prepare_schema_name}
\alias{prepare_schema_name}
\title{Parse and Sanitize a Connection Name into a Schema Name}
\usage{
prepare_schema_name(connection_name)
}
\arguments{
\item{connection_name}{The connection name.}
}
\value{
The schema name associated with the connection name.
}
\description{
Prepares a schema name from a connection name.  Currently a placeholder
for future logic to decouple connection names from schemas.
}
