% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{prepare_table_name}
\alias{prepare_table_name}
\title{Parse and Sanitize a Table Name}
\usage{
prepare_table_name(table_name)
}
\arguments{
\item{table_name}{The table name as a string.  Note that table names must
begin with an alpha character and only use alphanumerics, _, or the @@
sequence.  If @@ is found in a table name, it will be replaced by
the contents of the global variable called 'ENVIRONMENT' as long as that
variable only contains alpha characters.}
}
\value{
The prepared and sanitized table name.
}
\description{
Takes the given table name and prepares it for use in SQL.
}
