% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{replace_4ats}
\alias{replace_4ats}
\title{Replace @@ Within a String}
\usage{
replace_4ats(str, uppercase = FALSE)
}
\arguments{
\item{str}{String to search and replace.}

\item{uppercase}{Boolean to indicate if the name should be uppercased.
Defaults to FALSE.}
}
\value{
The prepared string.
}
\description{
What it says on the tin.  Will replace all ocurrences of @@
in the given string with the contents of the global variable named
'ENVIRONMENT' as long as that value contains only alpha characters.
}
