% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bettr_data.R
\name{rw_model}
\alias{rw_model}
\title{Random Walk based betting odds forecasting}
\usage{
rw_model(object,
         odds_vec,
         h = 36,
         has_na = FALSE
)
}
\arguments{
\item{object}{An object of class \code{bettr_data}.}

\item{odds_vec}{A numeric vector of odds.}

\item{h}{The number of minutes to be forecast.}

\item{has_na}{Should be true if there are NA values present in \code{object} and/or \code{odds_vec}. Note
that \code{predict.bettr_data} automatically checks whether NA values and/or gaps exist in the two and deals
with that accordingly.}
}
\value{
A list featuring the forecasted values, a forecast plot and forecasted values.
}
\description{
Random Walk based betting odds forecasting
}
\note{
It is highly recommended to use the \code{\link{predict.bettr_data}} function over the \code{\link{arima_model}},
\code{\link{ets_model}} and \code{\link{rw_model}} functions as it has checks for time gaps and NA values. Also note that the \code{predict.bettr_data}
function takes in the name of the odds column while \code{arima_model}, \code{ets_model}, \code{\link{skellam_model}} and \code{\link{rw_model}} take the vector
of the associated values.
}
\examples{
match <- subset(football, football$home_team == "Crystal Palace")
match <- tsibble::fill_gaps(make_bettr(match))
rw_model(match, match$draw_odds, h = 20, has_na = TRUE)
}
\seealso{
\code{\link{arima_model}}, \code{\link{ets_model}}, \code{\link{skellam_model}}
}
\author{
Ivan Cakic - <\email{ivan.cakic.2023@mumail.ie}>
}
