% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bettr_data.R
\name{skellam_model}
\alias{skellam_model}
\title{Skellam model forecasting future betting odds}
\usage{
skellam_model(object, odds_vec, h = 36, tickSize = 0.01, sims = 2000)
}
\arguments{
\item{object}{Object on which \code{skellam_model} is performed on.}

\item{odds_vec}{A vector of odds.}

\item{h}{The number of hours to be forecast. Defaults to 36 hours.}

\item{tickSize}{The minimum change in odds. Defaults to 0.01}

\item{sims}{The number of simulations}
}
\value{
A list featuring the forecasted values, the plot of the odds with respect to time and the model that was fitted
}
\description{
Skellam model forecasting future betting odds
}
\examples{
match <- subset(football, football$home_team == "Aston Villa")
match <- tsibble::fill_gaps(make_bettr(match))
ets_model(match, match$draw_odds, h = 18)
}
