% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{submit_task_report}
\alias{submit_task_report}
\title{Creates and Centrally Logs A Task Report}
\usage{
submit_task_report(bettr_task_key, last_report = NULL)
}
\arguments{
\item{last_report}{A dataframe, tibble, or list that can be
converted to JSON and sorted in the \code{last_report} field.}
}
\value{
A string in JSON format of the given report. Note
that parsing with \code{jsonlite::toJSON} may fail with complex
objects and JSON strings over 4000 characters are dropped
for exceeding storage limits.
}
\description{
The field BETTR_TASK.LAST_REPORT stores the last "report"
generated by the task R file, if any.  This is simply
the last returned object from the task R file, converted
to JSON.
This function is provided as a convenience for client code
to explicitly self-report to the task table and convert
the data to JSON for their own needs.
}
