% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{append_rows}
\alias{append_rows}
\title{Append Rows to a Table}
\usage{
append_rows(rows, connection_name, table_name)
}
\arguments{
\item{.}{a data frame, data.table, or tibble with rows of values to insert}

\item{conn}{The ROracle::connection object}

\item{name}{The name of the table}
}
\description{
Take the passed rows and append them to a table given a connection.
If the table does not exist, it is created with the given name.
Note that the schema for created tables is expected to be the database username.
}
