% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{append_rows}
\alias{append_rows}
\title{Append Rows to a Table}
\usage{
append_rows(rows, connection_name, table_name)
}
\arguments{
\item{rows}{a data frame, data.table, or tibble with rows of values to insert}

\item{connection_name}{Name of the database connection.  This is used to retrieve the}

\item{table_name}{The name of the table to receive data.}
}
\description{
Take the passed rows and append them to a table given a connection.
If the table does not exist, it is created with \code{ensure_table()}.
Note: the schema for created tables is expected to be the database username.
}
