% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bettr_data.R
\name{arima_model}
\alias{arima_model}
\title{Arima model forecasting future betting odds}
\usage{
arima_model(object, odds_vec, h = 36)
}
\arguments{
\item{object}{Object on which \code{arima_model} is performed on.}

\item{odds_vec}{A vector of odds.}

\item{h}{The number of hours to be forecast. Defaults to 36 hours.}
}
\value{
A list featuring the forecasted values, the plot of the odds with respect to time and the model that was fitted
}
\description{
Arima model forecasting future betting odds
}
\examples{
match <- subset(football, football$home_team == "Brentford")
match <- tsibble::fill_gaps(make_bettr(match))
arima_model(match, match$away_odds, h = 48)
}
