% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bettr-package.R
\docType{package}
\name{bettr-package}
\alias{bettr-package}
\title{bettr: a better way to explore what is best}
\description{
The bettr package provides a better way to explore what is best :)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/federicomarini/bettr}
  \item Report bugs at \url{https://github.com/federicomarini/bettr/issues}
}

}
\author{
Charlotte Soneson \email{charlottesoneson@gmail.com}

Federico Marini \email{marinif@uni-mainz.de}
}
\keyword{internal}
