% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bettr-package.R
\docType{package}
\name{bettr-package}
\alias{bettr-package}
\title{bettr: a better way to explore what is best}
\description{
The bettr package provides a better way to explore what is best :)
Details about how to use the package can be found in the vignette.
The main entry point is the \code{bettr()} function, which opens an
interactive application for exploring data consisting of multiple
parallel rankings of a set of entities (e.g., computational methods ranked
by their performance based on several different metrics).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/federicomarini/bettr}
  \item Report bugs at \url{https://github.com/federicomarini/bettr/issues}
}

}
\author{
Charlotte Soneson \email{charlottesoneson@gmail.com}

Federico Marini \email{marinif@uni-mainz.de}
}
\keyword{internal}
