\name{bettR-package}
\alias{bettR-package}
\alias{bettR}
\docType{package}
\title{
R extensions to be always loaded
}
\description{
Features that one always needs. General and not problem dependent. Good idea from other languages that fit here.
}
\details{
\tabular{ll}{
Package: \tab bettR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-05\cr
License: \tab What license is it under?\cr
}
}
\author{
Antonio Piccolboni <bettR@piccolboni.info}
\references{
Package \code{vadr} for pipes in R. Perl for \code{qw}
}

