% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETtR-package.R
\docType{package}
\name{bettr-package}
\alias{bettr}
\alias{bettr-package}
\title{BETtR:An R package for structuring, visualising, and exploring betting market odds data.}
\description{
This function serves as the package startup. Displaying an ASCII art as well as relevant information for package purpose, version history and other helpul tips
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/ThatObiGuy/bettr/issues}
}

}
\author{
Sean F. O'Leary - <\email{sean.oleary.2023@mumail.ie}>
}
\keyword{package}
