% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bettr-pkg.R
\docType{package}
\name{bettr-pkg}
\alias{bettr-pkg}
\title{bettr: a better way to explore what is best}
\description{
The bettr package provides a better way to explore what is best :)
}
\author{
Charlotte Soneson \email{charlottesoneson@gmail.com}

Federico Marini \email{marinif@uni-mainz.de}
}
