% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bettr-main.R
\name{bettr}
\alias{bettr}
\title{Launch bettr app to explore and aggregate performance metrics}
\usage{
bettr(
  df,
  methodCol = "Method",
  metricCol = "Metric",
  valueCol = "Value",
  initialWeights = NULL
)
}
\arguments{
\item{df}{A \code{data.frame} in long format. Must have at least
three columns: one providing the method ID (for the methods
that we want to compare), one providing the metric ID, and
one providing the value of each metric for each method.}

\item{methodCol, metricCol, valueCol}{Character scalars,
indicating which columns of \code{df} that contain method IDs, metric IDs
and metric values, respectively.}

\item{initialWeights}{Named numeric vector providing initial weights for
aggregating the metric scores. Must have one entry per metric included
in \code{df}.}
}
\description{
Launch bettr app to explore and aggregate performance metrics
}
\author{
Charlotte Soneson
}
