% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bettr-main.R
\name{bettr}
\alias{bettr}
\title{Launch bettr app to explore and aggregate performance metrics}
\usage{
bettr(
  df,
  methodCol = "Method",
  metrics = setdiff(colnames(df), methodCol),
  initialWeights = NULL
)
}
\arguments{
\item{df}{A \code{data.frame} in wide format.}

\item{methodCol}{Character scalar, indicating which column of \code{df}
that contains method IDs.}

\item{metrics}{Character vector, indicating which of the columns of
\code{df} that correspond to metrics of interest.}

\item{initialWeights}{Named numeric vector providing initial weights for
aggregating the metric scores. Must have one entry per metric included
in \code{df}.}
}
\description{
Launch bettr app to explore and aggregate performance metrics
}
\author{
Charlotte Soneson
}
