% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bettr-main.R
\name{bettr}
\alias{bettr}
\title{Launch bettr app to explore and aggregate performance metrics}
\usage{
bettr(
  df,
  idCol = "Method",
  metrics_num = setdiff(colnames(df), idCol),
  metrics_cat = c(),
  initialWeights = NULL,
  initialTransforms = list(),
  metricGroups = list(),
  bstheme = "darkly"
)
}
\arguments{
\item{df}{A \code{data.frame} in wide format.}

\item{idCol}{Character scalar, indicating which column of \code{df}
that contains IDs of the entities to be compared (e.g., methods).}

\item{metrics_num, metrics_cat}{Character vectors, indicating which of the
columns of \code{df} that correspond to, numeric and categorical,
respectively, metrics of interest.}

\item{initialWeights}{Named numeric vector providing initial weights for
aggregating the metric scores. Must have one entry per metric included
in \code{df}.}

\item{initialTransforms}{Named list with initial values of transformation
parameters for each metric.}

\item{metricGroups}{Named list of named character vectors. Each list entry
corresponds to one grouping of metrics. The grouping much be a named
vector indicating the respective group for each metric.}

\item{bstheme}{Character scalar giving the bootswatch theme for the app
(see https://bootswatch.com/). Default 'darkly'.}
}
\description{
Launch bettr app to explore and aggregate performance metrics
}
\author{
Charlotte Soneson
}
