% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{ensure_table}
\alias{ensure_table}
\title{Ensure a Table Exists for Appending Rows}
\usage{
ensure_table(rows, connection_name, table_name, skip_archive_table = FALSE)
}
\arguments{
\item{rows}{Table structure with representative data for the table to create/verify.}

\item{connection_name}{The snake_case name of the connection.}

\item{table_name}{The snake_case name of the table to ensure exists in the database.}

\item{skip_archive_table}{Whether to skip ensuring an identical "ARCH_" table.
Defaults to FALSE.}
}
\value{
Nothing
}
\description{
This function ensures a table exists in the named connection for appending rows.
Note that this function depends on a SQL file named \verb{create_table_<table_name>.sql}
in the connection-specific directory.
}
