% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{execute_stmts}
\alias{execute_stmts}
\title{Execute a Series of SQL Transactions}
\usage{
execute_stmts(connection_name, sql_file)
}
\arguments{
\item{connection_name}{The snake_case name of the connection.}

\item{sql_file}{the SQL part of the pattern: <SQL_DIR>/<CONNECTION_NAME>/\if{html}{\out{<SQL>}}.sql}
}
\value{
Nothing
}
\description{
Execute a Series of SQL Transactions
}
