% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{football}
\alias{football}
\title{Odds from bookkeeper Pinnacle - a series of English premier league matches taking place on the weekend starting 29/11/2025}
\format{
A data frame with 1573 observations (10 matches with ~160 observations from each) and 9 columns. It contains the required columns of a bettr_data object with additional info about maximum bettable amount and start times.
}
\usage{
data(football)
}
\description{
Data on the odds at different time points of a series of matches from the 2025 English premier league, including max bettable amounts at these intervals.
}
\examples{
data(football, package="bettr")
pairs(football[,-(1:2)], col=football$home_team)
}
\keyword{datasets}
