% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bettr.R
\name{make_bettr}
\alias{make_bettr}
\title{Title}
\usage{
make_bettr(
  data,
  event_id = "event_id",
  logged_time = "logged_time",
  home_odds = "home_odds",
  away_odds = "away_odds",
  draw_odds = "draw_odds",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
