% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBarPolarPlot.R
\name{makeBarPolarPlot}
\alias{makeBarPolarPlot}
\title{Create a bar/polar plot}
\usage{
makeBarPolarPlot(
  plotdata,
  scores,
  idCol,
  metricCol = "Metric",
  valueCol = "ScaledValue",
  weightCol = "Weight",
  scoreCol = "Score",
  metricGroupCol = "metricGroup",
  methods = NULL,
  labelSize = 10,
  showComposition = FALSE,
  scaleFactorPolars = 1,
  metricColors,
  metricCollapseGroup,
  metricGrouping
)
}
\arguments{
\item{plotdata}{A \code{data.frame} with columns representing methods,
metrics, scores, and weights. Typically obtained as
\code{prepData$plotdata}, where \code{prepData} is the output from
\code{bettrPrepare}.}

\item{scores}{A \code{data.frame} with columns representing methods,
aggregated scores, and any other method annotations. Typically
obtained as \code{prepData$scoredata}, where \code{prepData} is the
output from \code{bettrPrepare}.}

\item{idCol}{Character scalar indicating which column of \code{plotdata} and
\code{scores} contains the method IDs.}

\item{metricCol}{Character scalar indicating which column of \code{plotdata}
contains the metric IDs. Typically, \code{"Metric"}.}

\item{valueCol}{Character scalar indicating which column of \code{plotdata}
contains the metric values. Typically, \code{"ScaledValue"}.}

\item{weightCol}{Character scalar indicating which column of \code{plotdata}
contains the weight values. Typically, \code{"Weight"}.}

\item{scoreCol}{Character scalar indicating which column of \code{scores}
contains the aggregated score values. Typically, \code{"Score"}.}

\item{metricGroupCol}{Character scalar indicating which column of
\code{plotdata} contains the information about the metric group.
Typically, \code{"metricGroup"}.}

\item{methods}{Character vector containing the methods for which to make
polar plots. If \code{NULL} (default), all methods will be used.}

\item{labelSize}{Numeric scalar providing the size of the labels in the plot.}

\item{showComposition}{Logical scalar indicating whether to show the
composition of the score in the bar plots. This is only interpretable
if the scores are obtained via a weighted mean approach.}

\item{scaleFactorPolars}{Numeric scalar giving the scale factor determining
the size of the polar plots.}

\item{metricColors}{Named list with colors used for the metrics and
any other metric annotations. Typically obtained as
\code{prepData$metricColors}, where \code{prepData} is the output from
\code{bettrPrepare}.}

\item{metricCollapseGroup}{Logical scalar indicating whether metrics
should be collapsed by the group variable provided by
\code{metricGrouping}. Typically obtained as
\code{prepData$metricCollapseGroup}, where \code{prepData} is the
output from \code{bettrPrepare}.}

\item{metricGrouping}{Character scalar indicating the column of
\code{metricInfo} that was used to group metrics. Typically obtained as
\code{prepData$metricGrouping}, where \code{prepData} is the output
from \code{bettrPrepare}.}
}
\value{
A \code{ggplot} object.
}
\description{
Create a bar/polar plot. The input arguments for this functions are
typically generated using \code{\link{bettrPrepare}}, which ensures that
all required columns are available.
}
\examples{
## Generate example data
df <- data.frame(Method = c("M1", "M2", "M3"), 
                 metric1 = c(1, 2, 3),
                 metric2 = c(3, 1, 2))
metricInfo <- data.frame(Metric = c("metric1", "metric2", "metric3"),
                         Group = c("G1", "G2", "G2"))
idInfo <- data.frame(Method = c("M1", "M2", "M3"), 
                     Type = c("T1", "T1", "T2"))
prepData <- bettrPrepare(df = df, idCol = "Method", 
                         metricInfo = metricInfo, idInfo = idInfo)
makeBarPolarPlot(plotdata = prepData$plotdata, scores = prepData$scoredata, 
                 idCol = "Method", metricGroupCol = prepData$metricGroupCol,
                 metricColors = prepData$metricColors, 
                 metricCollapseGroup = prepData$metricCollapseGroup,
                 metricGrouping = prepData$metricGrouping, 
                 showComposition = TRUE)
                 
}
\author{
Charlotte Soneson
}
