% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePolarPlot.R
\name{makePolarPlot}
\alias{makePolarPlot}
\title{Create a polar plot}
\usage{
makePolarPlot(
  plotdata,
  idCol,
  metricCol = "Metric",
  valueCol = "ScaledValue",
  metricGroupCol = "metricGroup",
  labelSize = 10,
  metricColors,
  metricCollapseGroup,
  metricGrouping
)
}
\arguments{
\item{plotdata}{A \code{data.frame} with columns representing methods,
metrics, scores, and weights. Typically obtained as
\code{prepData$plotdata}, where \code{prepData} is the output from
\code{bettrPrepare}.}

\item{idCol}{Character scalar indicating which column of \code{plotdata} and
\code{scoredata} contains the method IDs.}

\item{metricCol}{Character scalar indicating which column of \code{plotdata}
contains the metric IDs. Typically, \code{"Metric"}.}

\item{valueCol}{Character scalar indicating which column of \code{plotdata}
contains the metric values. Typically, \code{"ScaledValue"}.}

\item{metricGroupCol}{Character scalar indicating which column of
\code{plotdata} contains the information about the metric group.
Typically, \code{"metricGroup"}.}

\item{labelSize}{Numeric scalar providing the size of the labels in the plot.}

\item{metricColors}{Named list with colors used for the metrics and
any other metric annotations. Typically obtained as
\code{prepData$metricColors}, where \code{prepData} is the output from
\code{bettrPrepare}.}

\item{metricCollapseGroup}{Logical scalar indicating whether metrics
should be collapsed by the group variable provided by
\code{metricGrouping}. Typically obtained as
\code{prepData$metricCollapseGroup}, where \code{prepData} is the
output from \code{bettrPrepare}.}

\item{metricGrouping}{Character scalar indicating the column of
\code{metricInfo} that was used to group metrics. Typically obtained as
\code{prepData$metricGrouping}, where \code{prepData} is the output
from \code{bettrPrepare}.}
}
\value{
A \code{ggplot} object.
}
\description{
Create a polar plot. The input arguments for this functions are
typically generated using \code{\link{bettrPrepare}}, which ensures that
all required columns are available.
}
\examples{
## Generate example data
df <- data.frame(Method = c("M1", "M2", "M3"), 
                 metric1 = c(1, 2, 3),
                 metric2 = c(3, 1, 2))
metricInfo <- data.frame(Metric = c("metric1", "metric2", "metric3"),
                         Group = c("G1", "G2", "G2"))
idInfo <- data.frame(Method = c("M1", "M2", "M3"), 
                     Type = c("T1", "T1", "T2"))
prepData <- bettrPrepare(df = df, idCol = "Method", 
                         metricInfo = metricInfo, idInfo = idInfo)
makePolarPlot(plotdata = prepData$plotdata, idCol = "Method", 
              metricGroupCol = prepData$metricGroupCol,
              metricColors = prepData$metricColors, 
              metricCollapseGroup = prepData$metricCollapseGroup,
              metricGrouping = prepData$metricGrouping)
                 
}
\author{
Charlotte Soneson
}
