% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bettr_data.R
\name{plot.bettr_data}
\alias{plot.bettr_data}
\title{Plot BETtR data (interactive)}
\usage{
\method{plot}{bettr_data}(x, odd = c("CF", "OF"), fixture = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{"bettr_data"} - You can use the inbuilt dataset,
\code{"football"} or using your own odds data after using the \code{\link{make_bettr}} function.}

\item{odd}{Either \code{"CF"} or \code{"OF"} for Closing Favorite or Opening Favorite respectively.
Defaults to \code{"CF"}.}

\item{fixture}{Optional additional column in \code{x} to use for tooltips
and legend instead of \code{event_id}. Defaults to NULL (uses \code{event_id}).}

\item{...}{Additional arguments passed to \code{ggiraph::girafe()} for specifying dimensions of output.}
}
\value{
a \code{ggiraph} object showing the odds movement over time
}
\description{
Plots percentage change of the odds of a (or many) opening favorite(s) change over time,
with interactive tooltips for the event IDs.
}
\examples{
football <- make_bettr(football, make_tsibble = FALSE)
plot(football)
plot(football, odd = "CF", width_svg = 8)

football2 <- football |>
  dplyr::mutate(fixture = paste(home_team, "vs", away_team))

plot(football2, fixture = "fixture")
}
\author{
Owen F. O'Connor - <\email{owen.oconnor.2024@mumail.ie}>
}
