% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bettr_data.R
\name{plot.bettr_data}
\alias{plot.bettr_data}
\title{Plot BETtR data}
\usage{
\method{plot}{bettr_data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"bettr_data"} - You can use the inbuilt dataset, \code{"football"} or using your own odds data after using the \code{link{make_bettr}} function.}

\item{...}{Catches unused arguments to \code{plot} (not currently implemented).}
}
\value{
a ggplot using geom_line to show the odds movement over time
}
\description{
Plots percentage change of the odds of a (or many) opening favorite(s) change over time.
}
\examples{
plot(football)
}
\author{
Owen F. O'Connor - <\email{owen.oconnor.2024@mumail.ie}>
}
