% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bettr_data.R
\name{predict.bettr_data}
\alias{predict.bettr_data}
\title{Title}
\usage{
\method{predict}{bettr_data}(data, h, model = c("all", "skellam", "arima", "ets"), ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
