% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bettr_data.R
\name{predict.bettr_data}
\alias{predict.bettr_data}
\title{Apply a list of models to a dataset featuring betting odds}
\usage{
predict(object,
       odds,
       h = 36,
       model = c("all", "arima", "ets", "skellam"),
       has_na = NULL,
       ...
)
}
\arguments{
\item{object}{An object of class \code{bettr_data}}

\item{odds}{A character string which is the name of the column the user wishes to investigate.}

\item{h}{The number of hours to be forecast.}

\item{model}{Models to be applied to the data.}

\item{has_na}{Should be true if there are NA values present in \code{object} and/or
if \code{object} has gaps in time data. If not specified, the function checks for gaps and
NAs and passes the result of the checks to the named functions.}

\item{...}{Extra parameters to be passed to the Skellam model.}
}
\value{
A list featuring the models fitted, their forecast plot and forecasted values.
}
\description{
A wrapper function for \code{\link{arima_model}}, \code{\link{ets_model}} and \code{\link{skellam_model}}
with some built-in checks for univariate time series and NA value handling necessary for arima and ets.
}
\examples{
data <- subset(football, home_team == "Sunderland")
data <- make_bettr(data)
predict(data, odds = "home_odds", h = 24, model = "all", sims = 1000, tickSize = 0.01)
}
\author{
Ivan Cakic - <\email{ivan.cakic.2023@mumail.ie}>
}
