% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{prepare_sql}
\alias{prepare_sql}
\title{Parse and Sanitize a SQL Literal}
\usage{
prepare_sql(sql)
}
\description{
Prepares a SQL file for execution with the ROracle library.
This only means that all occurrences of @@ in the SQL will be
replaced with the contents of the global variable named
'ENVIRONMENT' as long as that value contains only alpha characters.
}
