% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{resolve_rs_task_result}
\alias{resolve_rs_task_result}
\title{Resolve An Ended Task Session with the Bettr Host}
\usage{
resolve_rs_task_result(rs_task, bettr_task)
}
\arguments{
\item{rs_task}{A \code{callr::r_session} object where
our task R file has just stopped execution.}
}
\description{
Resolve An Ended Task Session with the Bettr Host
}
