% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{run_next_job_in_queue}
\alias{run_next_job_in_queue}
\title{Run Next Project Job in Better Queue}
\usage{
run_next_job_in_queue(
  project = Sys.getenv("BETTR_TASK_GIT_PROJECT"),
  branch = Sys.getenv("BETTR_TASK_GIT_BRANCH")
)
}
\arguments{
\item{project}{String name of the project, used to
filter for eligible jobs.}

\item{branch}{String name of the project branch, used
to filter for eligible jobs.}
}
\description{
This uses the provided git repository (project) and
branch name to query the host for the next runnable
job, being a whole group of tasks.
Note that this will attempt to rerun, in order, all
the tasks under the job, not just those expired,
errored, or not started.
This is the first job retrieved by this SQL:
\code{"{system.file(package = 'bettr')}/sql/BETTR_HOST/get_next_bettr_job.sql"}
}
