% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{run_next_task_in_queue}
\alias{run_next_task_in_queue}
\title{Run Next Project Task in Bettr Queue}
\usage{
run_next_task_in_queue(project)
}
\arguments{
\item{project}{String name of the project, used to
filter for eligible tasks.}
}
\description{
Run Next Project Task in Bettr Queue
}
